+++ src/com/kaylerrenslow/armaDialogCreator/gui/img/ADCImages.java
@@ -1,0 +1,14 @@
+package com.kaylerrenslow.armaDialogCreator.gui.img;
+
+import javafx.scene.image.Image;
+
+/**
+ Created by Kayler on 10/07/2016.
+ */
+public class ADCImages {
+	public static final Image ICON_ADC = new Image(ADCImagePaths.ICON_APP);
+	public static final Image ICON_FOLDER = new Image(ADCImagePaths.ICON_FOLDER);
+	public static final Image ICON_UNDO = new Image(ADCImagePaths.ICON_UNDO);
+	public static final Image ICON_REDO = new Image(ADCImagePaths.ICON_REDO);
+	public static final Image ICON_GEAR = new Image(ADCImagePaths.ICON_GEAR);
+}