+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/stringtable/StringTableEditorPopup.java
@@ -1,12 +1,10 @@
 package com.kaylerrenslow.armaDialogCreator.gui.main.stringtable;
 
-import com.kaylerrenslow.armaDialogCreator.arma.stringtable.Language;
-import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTable;
-import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableKey;
-import com.kaylerrenslow.armaDialogCreator.arma.stringtable.StringTableValue;
+import com.kaylerrenslow.armaDialogCreator.arma.stringtable.*;
+import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.DownArrowMenu;
 import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.SearchTextField;
 import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.SyntaxTextArea;
-import com.kaylerrenslow.armaDialogCreator.gui.img.Images;
+import com.kaylerrenslow.armaDialogCreator.gui.img.ADCImages;
 import com.kaylerrenslow.armaDialogCreator.gui.popup.StagePopup;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
 import com.kaylerrenslow.armaDialogCreator.main.Lang;
@@ -22,10 +20,11 @@ import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.geometry.Insets;
+import javafx.geometry.Pos;
 import javafx.scene.Cursor;
+import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.*;
 import javafx.scene.text.Font;
 import javafx.util.Callback;
@@ -55,37 +54,73 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 	}
 
 	private static class StringTableEditorTabPane extends TabPane {
+		private final ValueObserver<Language> previewLanguageObserver = new ValueObserver<>(KnownLanguage.Original);
+
 		public StringTableEditorTabPane(@NotNull StringTable table) {
-			getTabs().add(new EditTab(table));
-			getTabs().add(new ConfigTab(table));
+			getTabs().add(new EditTab(table, previewLanguageObserver));
+			getTabs().add(new ConfigTab(table, previewLanguageObserver));
 		}
 	}
 
 	private static class ConfigTab extends Tab { //set xml things like project name attribute (project=root tag of stringtable.xml)
-		public ConfigTab(@NotNull StringTable table) {
+
+		public ConfigTab(@NotNull StringTable table, @NotNull ValueObserver<Language> previewLanguageObserver) {
 			super(Lang.ApplicationBundle().getString("Popups.StringTable.Tab.Config.tab_title"));
-			setGraphic(new ImageView(Images.ICON_GEAR));
+			VBox root = new VBox(10);
+			root.setPadding(new Insets(10));
+			root.setFillWidth(true);
+			setContent(root);
+			setGraphic(new ImageView(ADCImages.ICON_GEAR));
 			setClosable(false);
 
 			ResourceBundle bundle = Lang.ApplicationBundle();
+
+			ComboBox<Language> comboBoxLanguage = new ComboBox<>(FXCollections.observableArrayList(KnownLanguage.values()));
+			comboBoxLanguage.getSelectionModel().select(previewLanguageObserver.getValue());
+			comboBoxLanguage.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<Language>() {
+				@Override
+				public void changed(ObservableValue<? extends Language> observable, Language oldValue, Language newValue) {
+					previewLanguageObserver.updateValue(newValue);
+				}
+			});
+			Label lblPreviewLanguage = new Label(bundle.getString("Popups.StringTable.Tab.Config.preview_language"), comboBoxLanguage);
+			lblPreviewLanguage.setContentDisplay(ContentDisplay.RIGHT);
+			root.getChildren().add(lblPreviewLanguage);
+
 		}
 	}
 
 	private static class EditTab extends Tab {
-		private final ObservableList<StringTableKeyDescriptor> listViewItems = FXCollections.observableList(new ArrayList<>(), new Callback<StringTableKeyDescriptor, javafx.beans
-				.Observable[]>() {
-
-			public javafx.beans.Observable[] call(StringTableKeyDescriptor param) {
-				return new javafx.beans.Observable[]{param.getKey().getValue().getLanguageTokenMap(), param.getKey().idObserver()};
-			}
-		}); //for some reason, can't have a LinkedList as the underlying list implementation if we want the list view to update the displayed cell text automatically
+		private final ObservableList<StringTableKeyDescriptor> listViewItems;
 
 		private final List<StringTableKeyDescriptor> allItems = new LinkedList<>();
 		private final ListView<StringTableKeyDescriptor> lvMatch = new ListView<>();
-		private final StringTableKeyEditorPane editorPane = new StringTableKeyEditorPane();
+		private final StringTableKeyEditorPane editorPane;
 
-		public EditTab(@NotNull StringTable table) {
+
+		public EditTab(@NotNull StringTable table, @NotNull ValueObserver<Language> previewLanguageObserver) {
 			super(Lang.ApplicationBundle().getString("Popups.StringTable.Tab.Edit.tab_title"));
+
+			listViewItems = FXCollections.observableList(new ArrayList<>(), new Callback<StringTableKeyDescriptor, javafx.beans
+					.Observable[]>() {
+
+				public javafx.beans.Observable[] call(StringTableKeyDescriptor param) {
+					return new javafx.beans.Observable[]{param.getKey().getValue().getLanguageTokenMap(), param.getKey().idObserver(), previewLanguageObserver};
+				}
+			}); //for some reason, can't have a LinkedList as the underlying list implementation if we want the list view to update the displayed cell text automatically
+
+			previewLanguageObserver.addListener(new ValueListener<Language>() {
+				@Override
+				public void valueUpdated(@NotNull ValueObserver<Language> observer, @Nullable Language oldValue, @Nullable Language newValue) {
+					for (StringTableKeyDescriptor descriptor : allItems) {
+						descriptor.setPreviewLanguage(newValue);
+					}
+				}
+			});
+
+			editorPane = new StringTableKeyEditorPane(previewLanguageObserver);
+
+
 			setClosable(false);
 			ResourceBundle bundle = Lang.ApplicationBundle();
 
@@ -95,6 +130,7 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			final String noContainerName = bundle.getString("Popups.StringTable.no_container");
 			for (StringTableKey key : table.getKeys()) {
 				StringTableKeyDescriptor descriptor = new StringTableKeyDescriptor(key, noPackageName, noContainerName);
+				descriptor.setPreviewLanguage(previewLanguageObserver.getValue());
 				allItems.add(descriptor);
 				listViewItems.add(descriptor);
 			}
@@ -133,18 +169,46 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 						String searchTextAfter = searchText.substring(1);
 						switch (searchText.charAt(0)) {
 							case '\'': //intentional fall through
-							case '"': {
+							case '"': { //search by text content
 								searchTextAfter = searchTextAfter.toLowerCase();
-								for (StringTableKeyDescriptor key : allItems) {
-									for (Map.Entry<Language, String> entry : key.getKey().getValue().getLanguageTokenMap().entrySet()) {
+								for (StringTableKeyDescriptor descriptor : allItems) {
+									for (Map.Entry<Language, String> entry : descriptor.getKey().getValue().getLanguageTokenMap().entrySet()) {
 										if (entry.getValue().toLowerCase().contains(searchTextAfter)) {
-											lvMatch.getItems().add(key);
+											lvMatch.getItems().add(descriptor);
+											break;
 										}
 									}
 								}
 								break;
 							}
-							case '!': {
+							case '!': { //check if key does **not** have language
+								searchTextAfter = searchTextAfter.toLowerCase();
+								for (StringTableKeyDescriptor descriptor : allItems) {
+									boolean matched = false;
+									for (Language lang : descriptor.getKey().getValue().getLanguageTokenMap().keySet()) {
+										if (lang.getName().toLowerCase().contains(searchTextAfter)) {
+											matched = true;
+											break;
+										}
+									}
+									if (!matched) {
+										lvMatch.getItems().add(descriptor);
+									}
+								}
+								break;
+							}
+							case '+': { //check if key **has** a language
+								searchTextAfter = searchTextAfter.toLowerCase();
+								for (StringTableKeyDescriptor descriptor : allItems) {
+									for (Language lang : descriptor.getKey().getValue().getLanguageTokenMap().keySet()) {
+										if (lang.getName().toLowerCase().contains(searchTextAfter)) {
+											lvMatch.getItems().add(descriptor);
+											break;
+										}
+									}
+								}
+							}
+							case '/': { //search by package name and container name
 								String searchPackageName;
 								String searchContainerName = "";
 								if (searchTextAfter.contains("/")) {
@@ -208,20 +272,25 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			setContent(vbRoot);
 
 		}
+
+
 	}
 
 
 	private static class StringTableKeyEditorPane extends StackPane {
 		private static final Font FONT_KEY_ID = Font.font(15);
-		private static final String STR_ = "str_";
 
-		private final LanguageSelectionPane languagePane = new LanguageSelectionPane();
+		private final LanguageSelectionPane languagePane;
 		private final String noKeySelected = Lang.ApplicationBundle().getString("Popups.StringTable.Editor.no_selected_key");
 		private final TextField tfKeyId = new TextField();
-		private final Label lblStr_ = new Label();
+		private final Menu menuAddLanguage;
+		private final Menu menuRemoveLanguage;
 		private StringTableKey key;
 
-		public StringTableKeyEditorPane() {
+		public StringTableKeyEditorPane(@NotNull ValueObserver<Language> previewLanguageObserver) {
+			languagePane = new LanguageSelectionPane(previewLanguageObserver);
+			ResourceBundle bundle = Lang.ApplicationBundle();
+
 			HBox paneContent = new HBox(10);
 			getChildren().add(paneContent);
 
@@ -236,22 +305,25 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 						return;
 					}
 					newValue = newValue != null ? newValue : "";
-					key.setId(STR_ + newValue);
+					key.setId(newValue);
+				}
+			});
+			HBox.setHgrow(tfKeyId, Priority.ALWAYS);
+
+			menuAddLanguage = new Menu(bundle.getString("Popups.StringTable.add_language"));
+			menuRemoveLanguage = new Menu(bundle.getString("Popups.StringTable.remove_language"));
+
+			MenuItem miDefineLanguage = new MenuItem(bundle.getString("Popups.StringTable.define_language"));
+			miDefineLanguage.setOnAction(new EventHandler<ActionEvent>() {
+				@Override
+				public void handle(ActionEvent event) {
+
 				}
 			});
 
-			lblStr_.setFont(FONT_KEY_ID);
-			lblStr_.setCursor(Cursor.TEXT);
-			lblStr_.setOnMousePressed(new EventHandler<MouseEvent>() {
-				@Override
-				public void handle(MouseEvent event) {
-					event.consume();
-					tfKeyId.positionCaret(0);
-					tfKeyId.requestFocus();
-					tfKeyId.deselect();
-				}
-			});
-			VBox vboxLeft = new VBox(10, new HBox(lblStr_, tfKeyId), languagePane);
+			HBox hboxKey = new HBox(5, new DownArrowMenu(menuAddLanguage, menuRemoveLanguage/*, new SeparatorMenuItem(), miDefineLanguage*/), tfKeyId);
+			hboxKey.setAlignment(Pos.CENTER_LEFT);
+			VBox vboxLeft = new VBox(10, hboxKey, languagePane);
 			vboxLeft.setFillWidth(true);
 
 			SyntaxTextArea textArea = new SyntaxTextArea();
@@ -266,7 +338,7 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 				public void changed(ObservableValue<? extends String> observable, String oldValue, String newValue) {
 					if (key != null) {
 						newValue = newValue != null ? newValue : "";
-						Language language = languagePane.getValueObserver().getValue();
+						Language language = languagePane.getChosenLanguageObserver().getValue();
 						if (language != null) {
 							key.getValue().getLanguageTokenMap().replace(language, newValue);
 						}
@@ -274,15 +346,24 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 				}
 			});
 			textArea.getStyleClass().add("bordered-syntax-text-area");
-			languagePane.getValueObserver().addValueListener(new ReadOnlyValueListener<Language>() {
+			languagePane.getChosenLanguageObserver().addValueListener(new ReadOnlyValueListener<Language>() {
 				@Override
 				public void valueUpdated(@NotNull ReadOnlyValueObserver<Language> observer, @Nullable Language oldValue, @Nullable Language selected) {
 					if (key == null || selected == null) {
 						textArea.replaceText("");
 					} else {
-						textArea.replaceText(key.getValue().getLanguageTokenMap().get(selected));
+						Map.Entry<Language, String> entry = key.getValue().getFirstLanguageTokenEntry();
+						String s;
+						if (entry == null || entry.getValue() == null) {
+							s = "";
+						} else {
+							s = entry.getValue();
+						}
+
+						textArea.replaceText(s);
 					}
 				}
+
 			});
 
 			paneContent.getChildren().addAll(vboxLeft, textArea);
@@ -292,47 +373,118 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			setKey(null);
 		}
 
+
+		private static final Pattern PATTERN = Pattern.compile("(%[0-9]+)");
+
+		private static StyleSpans<Collection<String>> computeHighlighting(String text) {
+			text = text.replaceAll("%%", "__");//prevent matching %%1 or %%%%1. We can't remove the %% however because that would mess with the indexes
+			Matcher matcher = PATTERN.matcher(text);
+			int lastKwEnd = 0;
+			StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
+			while (matcher.find()) {
+				spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
+				spansBuilder.add(Collections.singleton("param"), matcher.end() - matcher.start());
+				lastKwEnd = matcher.end();
+			}
+			spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
+			return spansBuilder.create();
+		}
+
+
 		public void setKey(@Nullable StringTableKey key) {
 			this.key = key;
 			languagePane.setToKey(key);
 			if (key == null) {
-				lblStr_.setText(noKeySelected);
-				tfKeyId.setText("");
+				tfKeyId.setText(noKeySelected);
 			} else {
-				lblStr_.setText(STR_);
-				tfKeyId.setText(key.getIdWithoutStr_());
+				tfKeyId.setText(key.getId());
 			}
+			handleKeyDownMenus();
 			setDisable(key == null);
 		}
-	}
 
-	private static final Pattern PATTERN = Pattern.compile("(%[0-9]+)");
+		private void handleKeyDownMenus() {
+			menuRemoveLanguage.getItems().clear();
+			menuAddLanguage.getItems().clear();
+			if (key == null) {
+				return;
+			}
+			for (Language language : KnownLanguage.values()) {
+				boolean found = false;
+				MenuItem miLanguage = new MenuItem(language.getName());
+				miLanguage.setUserData(language);
 
-	private static StyleSpans<Collection<String>> computeHighlighting(String text) {
-		text = text.replaceAll("%%", "__");//prevent matching %%1 or %%%%1. We can't remove the %% however because that would mess with the indexes
-		Matcher matcher = PATTERN.matcher(text);
-		int lastKwEnd = 0;
-		StyleSpansBuilder<Collection<String>> spansBuilder = new StyleSpansBuilder<>();
-		while (matcher.find()) {
-			spansBuilder.add(Collections.emptyList(), matcher.start() - lastKwEnd);
-			spansBuilder.add(Collections.singleton("param"), matcher.end() - matcher.start());
-			lastKwEnd = matcher.end();
+				for (Map.Entry<Language, String> entry : key.getValue().getLanguageTokenMap().entrySet()) {
+					if (entry.getKey().equals(language)) {
+						found = true;
+						menuRemoveLanguage.getItems().add(miLanguage);
+						miLanguage.setOnAction(new LanguageMenuItemEvent(true, key, language, miLanguage, menuAddLanguage, menuRemoveLanguage));
+						break;
+					}
+				}
+				if (!found) {
+					menuAddLanguage.getItems().add(miLanguage);
+					miLanguage.setOnAction(new LanguageMenuItemEvent(false, key, language, miLanguage, menuAddLanguage, menuRemoveLanguage));
+				}
+			}
 		}
-		spansBuilder.add(Collections.emptyList(), text.length() - lastKwEnd);
-		return spansBuilder.create();
+
+		private class LanguageMenuItemEvent implements EventHandler<ActionEvent> {
+			private final StringTableKey key;
+			private final Language language;
+			private final MenuItem miLanguage;
+			private final Menu menuAddLanguage;
+			private final Menu menuRemoveLanguage;
+			private boolean added;
+
+			public LanguageMenuItemEvent(boolean added, @NotNull StringTableKey key, @NotNull Language language, @NotNull MenuItem miLanguage, @NotNull Menu menuAddLanguage, @NotNull Menu
+					menuRemoveLanguage) {
+				this.added = added;
+				this.key = key;
+				this.language = language;
+				this.miLanguage = miLanguage;
+				this.menuAddLanguage = menuAddLanguage;
+				this.menuRemoveLanguage = menuRemoveLanguage;
+			}
+
+			@Override
+			public void handle(ActionEvent event) {
+				added = !added;
+				if (added) {
+					menuAddLanguage.getItems().remove(miLanguage);
+					menuRemoveLanguage.getItems().add(miLanguage);
+					key.getValue().getLanguageTokenMap().put(language, "");
+				} else {
+					menuRemoveLanguage.getItems().remove(miLanguage);
+					menuAddLanguage.getItems().add(miLanguage);
+					key.getValue().getLanguageTokenMap().remove(language);
+				}
+
+			}
+		}
 	}
 
 	private static class LanguageSelectionPane extends FlowPane {
 
 		private static final String SELECTED_LINK_STYLE = "-fx-text-fill:darkgreen;-fx-font-weight:bold;";
+		private static final Comparator<Node> hyperlinkComparator = new Comparator<Node>() {
+			@Override
+			public int compare(Node o1, Node o2) {
+				Hyperlink h1 = (Hyperlink) o1;
+				Hyperlink h2 = (Hyperlink) o2;
+				return h1.getText().compareTo(h2.getText());
+			}
+		};
 
-		private final ValueObserver<Language> valueObserver = new ValueObserver<>(null);
+		private final ValueObserver<Language> chosenLanguageObserver = new ValueObserver<>(null);
 		private final List<Hyperlink> links = new LinkedList<>();
+		private ValueObserver<Language> previewLanguageObserver;
 
-		public LanguageSelectionPane() {
+		public LanguageSelectionPane(ValueObserver<Language> previewLanguageObserver) {
 			super(10, 10);
+			this.previewLanguageObserver = previewLanguageObserver;
 			setToKey(null);
-			valueObserver.addListener(new ValueListener<Language>() {
+			chosenLanguageObserver.addListener(new ValueListener<Language>() {
 				@Override
 				public void valueUpdated(@NotNull ValueObserver<Language> observer, @Nullable Language oldValue, @Nullable Language newValue) {
 					for (Hyperlink hyperlink : links) {
@@ -352,7 +504,7 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			setDisable(key == null);
 
 			//force an update to null so that if the Language is the same as the previous in this observer, the listeners will get notified below at comment marked with: ****
-			valueObserver.updateValue(null);
+			chosenLanguageObserver.updateValue(null);
 
 			if (key == null) {
 				return;
@@ -371,15 +523,11 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 					}
 				}
 			});
-			Language setToMe = null;
 			for (Map.Entry<Language, String> token : tableValue.getLanguageTokenMap().entrySet()) {
 				addLanguage(token.getKey());
-				if (setToMe == null) {
-					setToMe = token.getKey();
-				}
 			}
 
-			valueObserver.updateValue(setToMe); //**** (see above comment with valueObserver.updateValue(null))
+			chosenLanguageObserver.updateValue(previewLanguageObserver.getValue()); //**** (see above comment with valueObserver.updateValue(null))
 		}
 
 		private void addLanguage(@NotNull Language language) {
@@ -388,16 +536,18 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 			}
 			Hyperlink hyperlinkLanguage = new Hyperlink(language.getName());
 			links.add(hyperlinkLanguage);
+			links.sort(hyperlinkComparator);
 			hyperlinkLanguage.setUserData(language);
 			hyperlinkLanguage.setOnAction(new EventHandler<ActionEvent>() {
 				@Override
 				public void handle(ActionEvent event) {
 					hyperlinkLanguage.setVisited(false);
-					valueObserver.updateValue(language);
+					chosenLanguageObserver.updateValue(language);
 				}
 			});
 
-			getChildren().add(hyperlinkLanguage);
+			getChildren().clear();
+			getChildren().addAll(links);
 		}
 
 		private void removeLanguage(@NotNull Language language) {
@@ -424,14 +574,14 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 		}
 
 		@NotNull
-		public ReadOnlyValueObserver<Language> getValueObserver() {
-			return valueObserver.getReadOnlyValueObserver();
+		public ReadOnlyValueObserver<Language> getChosenLanguageObserver() {
+			return chosenLanguageObserver.getReadOnlyValueObserver();
 		}
+
 	}
 
 	private static class StringTableKeyDescriptor {
-		private static final int MAX_TEXT_LENGTH = 40;
-		private static final String FORMAT = "%-40s %-" + (MAX_TEXT_LENGTH + 2 + 3 /*+2 to for quotes on either side of text, +3 for ...*/) + "s (%s / %s)";
+		private static final String FORMAT = "%-40s (%20s / %-20s) %1s %s";
 
 		private StringTableKey key;
 		private final String noPackageName;
@@ -454,18 +604,31 @@ public class StringTableEditorPopup extends StagePopup<VBox> {
 		@Override
 		public String toString() {
 			String text = "";
+			boolean defaultMatch = true;
 			if (language != null) {
-				text = key.getValue().getLanguageTokenMap().get(language).replaceAll("[\r\n\t]", "");
+				text = key.getValue().getLanguageTokenMap().get(language);
+				if (text == null) {
+					defaultMatch = false;
+					Map.Entry<Language, String> entry = key.getValue().getFirstLanguageTokenEntry();
+					if (entry == null || entry.getValue() == null) {
+						text = "";
+					} else {
+						text = entry.getValue();
+					}
+				}
 				if (text == null) {
 					text = "";
+				} else {
+					text = text.replaceAll("[\r\n\t]", "");
 				}
 			}
 			return String.format(
 					FORMAT,
 					key.getId(),
-					"\"" + text.substring(0, Math.min(MAX_TEXT_LENGTH, text.length())) + (text.length() > MAX_TEXT_LENGTH ? "..." : "") + "\"",
 					key.getPackageName() != null ? key.getPackageName() : noPackageName,
-					key.getContainerName() != null ? key.getContainerName() : noContainerName
+					key.getContainerName() != null ? key.getContainerName() : noContainerName,
+					defaultMatch ? "" : "!",
+					"\"" + text + "\""
 			);
 		}